♻️  Refactor: extract common component
Add generic BackLinkRow component
Move RadioList components to common
