+++ tenantv3/src/components/account/TenantPanel.vue
@@ -80,7 +80,7 @@ import { User } from 'df-shared-next/src/models/User'
 import { DfDocument } from 'df-shared-next/src/models/DfDocument'
 import { AnalyticsService } from '../../services/AnalyticsService'
 import GuarantorsSection from '../../components/account/GuarantorsSection.vue'
-import RowListItem from '../../components/documents/RowListItem.vue'
+import RowListItem from '../documents/RowListItem.vue'
 import FileRowListItem from '../../components/documents/FileRowListItem.vue'
 import { DocumentTypeConstants } from '../../components/documents/share/DocumentTypeConstants'
 import { useRouter } from 'vue-router'
