+++ tenantv3/src/components/residency/ChooseResidency.vue
@@ -34,9 +34,9 @@
 </template>
 
 <script setup lang="ts">
-import RadioList from './lib/RadioList.vue'
+import RadioList from '../common/RadioList.vue'
 import DfButton from 'df-shared-next/src/Button/DfButton.vue'
-import RadioListItem from './lib/RadioListItem.vue'
+import RadioListItem from '../common/RadioListItem.vue'
 import ResidencyFooter from './lib/ResidencyFooter.vue'
 import { useI18n } from 'vue-i18n'
 import { AnalyticsService } from '@/services/AnalyticsService'
