+++ tenantv3/src/components/residency/lib/BackLinkRow.vue
@@ -1,11 +1,5 @@
 <template>
-  <p class="display--flex align-items--center">
-    <RiCheckboxCircleLine color="var(--primary)" size="20px" class="fr-mr-1w no-shrink" />
-    <span>{{ label }}</span>
-    <DfButton tertiary-no-outline class="fr-ml-auto" @click="onClick">
-      {{ t('edit') }} <RiArrowGoBackLine size="1rem" class="fr-ml-1w" />
-    </DfButton>
-
+  <BackLinkRow :label="label" @click="onClick">
     <ModalComponent v-if="showChangeSituation" @close="ignore">
       <template #body>
         <i18n-t keypath="delete-notice" tag="p">
@@ -23,12 +17,11 @@
         </ul>
       </template>
     </ModalComponent>
-  </p>
+  </BackLinkRow>
 </template>
 
 <script setup lang="ts">
 import { useRouter, type RouterLinkProps } from 'vue-router'
-import { RiArrowGoBackLine, RiCheckboxCircleLine } from '@remixicon/vue'
 import { useI18n } from 'vue-i18n'
 import { ref } from 'vue'
 import DfButton from 'df-shared-next/src/Button/DfButton.vue'
@@ -36,6 +29,7 @@ import ModalComponent from 'df-shared-next/src/components/ModalComponent.vue'
 import { useTenantStore } from '@/stores/tenant-store'
 import { AnalyticsService } from '@/services/AnalyticsService'
 import { useResidencyState } from '../residencyState'
+import BackLinkRow from '@/components/common/BackLinkRow.vue'
 
 const props = defineProps<{ label: string; to: RouterLinkProps['to'] }>()
 const emit = defineEmits<{ edit: [] }>()
@@ -71,23 +65,15 @@ const ignore = () => {
 }
 </script>
 
-<style scoped>
-.no-shrink {
-  flex-shrink: 0;
-}
-</style>
-
 <i18n>
 {
   "en": {
-    "edit": "Edit",
     "cancel": "Cancel",
     "change-situation": "Change your situation",
     "delete-warning": "if you change your situation, your supporting documents will be deleted",
     "delete-notice": "Please note: {0}. You will need to provide new supporting documents."
   },
   "fr": {
-    "edit": "Modifier",
     "cancel": "Annuler",
     "change-situation": "Modifier votre situation",
     "delete-warning": "modifier votre situation entraîne la suppression de vos justificatifs",
